Change type to double

+++ src/main/java/id/ac/ui/cs/advprog/tutorial4/core/code/GenerateCodeFactory.java
@@ -15,11 +15,11 @@ public class GenerateCodeFactory {
     public RedeemCode generateCode(String itemType, String codeType, String code, String itemName, Map<String, String> data) {
         code = randomCode(code);
         if (itemType.equals("merch") && codeType.equals("promo")) {
-            return new PromoMerchCode(code, itemName, Integer.parseInt(data.get("discount")));
+            return new PromoMerchCode(code, itemName, Double.parseDouble(data.get("discount")));
         } else if (itemType.equals("merch") && codeType.equals("giveaway")) {
             return new GiveawayMerchCode(code, itemName, Integer.parseInt(data.get("amount")));
         } else if (itemType.equals("membership") && codeType.equals("promo")) {
-            return new PromoMembershipCode(code, itemName, Integer.parseInt(data.get("discount")));
+            return new PromoMembershipCode(code, itemName, Double.parseDouble(data.get("discount")));
         } else if (itemType.equals("membership") && codeType.equals("giveaway")) {
             return new GiveawayMembershipCode(code, itemName, Integer.parseInt(data.get("amount")));
         } else {
